backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pillow==10.2.0
razorpay==1.4.1
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
jinja2==3.1.3
email-validator==2.1.0
httpx==0.26.0
alembic==1.13.1
